AioCb::Drop will now panic for in-progress AioCb
Printing a warning message to stderr isn't really appropriate, because
there's no way to guarantee that stderr is even valid.  Nor is
aio_suspend necessarily an appropriate action to take.
